@@ -1,10 +1,10 @@
 name: coast
 description: PageView wrapper that supports Hero-like animations
-version: 1.0.2
+version: 2.0.1
 repository: https://github.com/spkersten/flutter_coast
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -14,3 +14,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_test_ui:
