@@ -1,13 +1,14 @@
 import 'dart:math';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 
 /// A wrapper of [PageView] that support observation of transitions between pages.
 class Coast extends StatefulWidget {
   const Coast({
-    @required this.beaches,
-    @required this.controller,
+    required this.beaches,
+    required this.controller,
     this.observers,
     this.physics,
     this.allowImplicitScrolling = false,
@@ -17,33 +18,55 @@ class Coast extends StatefulWidget {
     this.onPageChanged,
     this.dragStartBehavior = DragStartBehavior.start,
     this.clipBehavior = Clip.hardEdge,
-    Key key,
-  })  : assert(beaches != null),
-        assert(beaches.length > 0),
-        assert(controller != null),
-        assert(allowImplicitScrolling != null),
+    Key? key,
+  })  : assert(beaches.length > 0),
         super(key: key);
 
   final List<Beach> beaches;
-  final List<CoastObserver> observers;
+  final List<CoastObserver>? observers;
   final CoastController controller;
 
   /// See [PageView]
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
   final bool allowImplicitScrolling;
-  final String restorationId;
+  final String? restorationId;
   final Axis scrollDirection;
   final bool reverse;
-  final ValueChanged<int> onPageChanged;
+  final ValueChanged<int>? onPageChanged;
   final DragStartBehavior dragStartBehavior;
   final Clip clipBehavior;
 
   @override
   State<StatefulWidget> createState() => CoastState();
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(IterableProperty<Beach>('beaches', beaches))
+      ..add(IterableProperty<CoastObserver>('observers', observers))
+      ..add(DiagnosticsProperty<CoastController>('controller', controller))
+      ..add(DiagnosticsProperty<ScrollPhysics?>('physics', physics))
+      ..add(DiagnosticsProperty<bool>(
+          'allowImplicitScrolling', allowImplicitScrolling))
+      ..add(StringProperty('restorationId', restorationId))
+      ..add(EnumProperty<Axis>('scrollDirection', scrollDirection))
+      ..add(DiagnosticsProperty<bool>('reverse', reverse))
+      ..add(ObjectFlagProperty<ValueChanged<int>?>.has(
+          'onPageChanged', onPageChanged))
+      ..add(EnumProperty<DragStartBehavior>(
+          'dragStartBehavior', dragStartBehavior))
+      ..add(EnumProperty<Clip>('clipBehavior', clipBehavior));
+  }
 }
 
 class CoastController {
-  double get beach => _pageController.page;
+  CoastController({int initialPage = 0})
+      : _pageController = PageController(initialPage: initialPage);
+
+  final PageController _pageController;
+
+  double? get beach => _pageController.page;
 
   void dispose() {
     _pageController.dispose();
@@ -52,29 +75,27 @@ class CoastController {
   PageController get instance => _pageController;
 
   Future<void> animateTo({
-    @required int beach,
+    required int beach,
     Duration duration = const Duration(milliseconds: 200),
     Curve curve = Curves.fastOutSlowIn,
   }) async {
     await _pageController.animateToPage(beach,
         duration: duration, curve: curve);
   }
-
-  final _pageController = PageController(keepPage: true);
 }
 
 class CoastState extends State<Coast> {
-  double _previousOffset;
-  TransitionAnimation progress;
-  int _sourcePage;
-  int _targetPage;
+  late double _previousOffset;
+  TransitionAnimation? progress;
+  int _sourcePage = 0;
+  int? _targetPage;
 
   final _overlayKey =
       GlobalKey<OverlayState>(debugLabel: "CoastState's Overlay");
 
-  OverlayState get overlay => _overlayKey.currentState;
+  OverlayState? get overlay => _overlayKey.currentState;
 
-  PageController get pageController => widget.controller._pageController;
+  PageController get _pageController => widget.controller._pageController;
 
   double _round(double value, int precision) {
     final f = pow(10, precision);
@@ -85,97 +106,101 @@ class CoastState extends State<Coast> {
   void initState() {
     super.initState();
 
-    _sourcePage = 0;
+    _sourcePage = _pageController.initialPage;
     _previousOffset = 0.0;
 
-    pageController.addListener(() {
-      if (context != null) {
-        // Get rid of over-scrolling
-        final offset = _round(
-            pageController.page
-                .clamp(0.0, widget.beaches.length - 1)
-                .toDouble(),
-            6);
-        if (offset == _previousOffset) return;
-
-        // Determine between which two pages we are scrolling
-        final newSourcePage =
-            calculateNewSourcePage(offset: offset, sourcePage: _sourcePage);
-
-        final newTargetPage = calculateNewTargetPage(
-            offset: offset, newSourcePage: newSourcePage);
-
-        if (shouldFinishTransition(
-            newTargetPage: newTargetPage, newSourcePage: newSourcePage)) {
-          // Finish previous transition
-          progress
-            ..value = (newSourcePage < _sourcePage) ? 0.0 : 1.0
-            ..dispose();
-          progress = null;
-          _targetPage = null;
-        }
-
-        _sourcePage = newSourcePage;
-
-        if (shouldStartNewTransition(newTargetPage: newTargetPage)) {
-          // Start new a transition
-          _targetPage = newTargetPage;
-
-          final direction = _targetPage > _sourcePage
-              ? BeachTransitionDirection.right
-              : BeachTransitionDirection.left;
-
-          progress = TransitionAnimation();
-
-          for (final observer in widget.observers ?? <CoastObserver>[]) {
-            observer.coast = this;
-            observer.startTransition(widget.beaches[_targetPage],
-                widget.beaches[_sourcePage], direction, progress);
-          }
-        }
-
-        // Update progress of the transition that is in progress
-        if (progress != null) {
-          if (_targetPage > _sourcePage)
-            progress.value = offset - _sourcePage;
-          else
-            progress.value = _sourcePage - offset;
-        }
-
-        _previousOffset = offset;
+    _pageController.addListener(() {
+      // Get rid of over-scrolling
+      final offset = _round(
+          _pageController.page!
+              .clamp(0.0, widget.beaches.length - 1)
+              .toDouble(),
+          6);
+      if (offset == _previousOffset) {
+        return;
       }
+
+      // Determine between which two pages we are scrolling
+      final newSourcePage =
+          calculateNewSourcePage(offset: offset, sourcePage: _sourcePage);
+
+      final newTargetPage =
+          calculateNewTargetPage(offset: offset, newSourcePage: newSourcePage);
+
+      if (shouldFinishTransition(
+          newTargetPage: newTargetPage, newSourcePage: newSourcePage)) {
+        // Finish previous transition
+        progress
+          ?..value = (newSourcePage < _sourcePage) ? 0.0 : 1.0
+          ..dispose();
+        progress = null;
+        _targetPage = null;
+      }
+
+      _sourcePage = newSourcePage;
+
+      if (shouldStartNewTransition(newTargetPage: newTargetPage)) {
+        // Start new a transition
+        _targetPage = newTargetPage;
+
+        final direction = _targetPage! > _sourcePage
+            ? BeachTransitionDirection.right
+            : BeachTransitionDirection.left;
+
+        progress = TransitionAnimation();
+
+        for (final observer in widget.observers ?? <CoastObserver>[]) {
+          observer
+            ..coast = this
+            ..startTransition(widget.beaches[_targetPage!],
+                widget.beaches[_sourcePage], direction, progress);
+        }
+      }
+
+      // Update progress of the transition that is in progress
+      if (progress != null) {
+        if (_targetPage! > _sourcePage) {
+          progress!.value = offset - _sourcePage;
+        } else {
+          progress!.value = _sourcePage - offset;
+        }
+      }
+
+      _previousOffset = offset;
     });
   }
 
   @visibleForTesting
   int calculateNewSourcePage(
-      {@required double offset, @required int sourcePage}) {
-    if (offset >= (sourcePage + 1))
+      {required double offset, required int sourcePage}) {
+    if (offset >= (sourcePage + 1)) {
       return sourcePage + (offset - sourcePage).floor();
-    else if (offset <= (sourcePage - 1))
+    } else if (offset <= (sourcePage - 1)) {
       return sourcePage - (sourcePage - offset).floor();
-    else
+    } else {
       return sourcePage;
+    }
   }
 
   @visibleForTesting
-  int calculateNewTargetPage(
-      {@required double offset, @required int newSourcePage}) {
-    if (offset > newSourcePage)
+  int? calculateNewTargetPage(
+      {required double offset, required int newSourcePage}) {
+    if (offset > newSourcePage) {
       return newSourcePage + 1;
-    else if (offset < newSourcePage)
+    } else if (offset < newSourcePage) {
       return newSourcePage - 1;
-    else
-      return null; // ignore: avoid_returning_null
+    } else {
+      return null;
+    }
   }
 
   @visibleForTesting
-  bool shouldFinishTransition({int newTargetPage, int newSourcePage}) =>
+  bool shouldFinishTransition({int? newTargetPage, int? newSourcePage}) =>
       progress != null &&
       (_targetPage != newTargetPage || _sourcePage != newSourcePage);
 
   @visibleForTesting
-  bool shouldStartNewTransition({int newTargetPage}) =>
+  bool shouldStartNewTransition({int? newTargetPage}) =>
       progress == null && newTargetPage != null;
 
   @override
@@ -187,10 +212,8 @@ class CoastState extends State<Coast> {
             Temporary fix to round the width till the issue is fixed */
             width: constraints.maxWidth.roundToDouble(),
             child: PageView(
-              controller: pageController,
+              controller: _pageController,
               physics: widget.physics,
-              children:
-                  widget.beaches.map((beach) => beach.build(context)).toList(),
               allowImplicitScrolling: widget.allowImplicitScrolling,
               restorationId: widget.restorationId,
               scrollDirection: widget.scrollDirection,
@@ -199,28 +222,42 @@ class CoastState extends State<Coast> {
               onPageChanged: widget.onPageChanged,
               dragStartBehavior: widget.dragStartBehavior,
               clipBehavior: widget.clipBehavior,
+              children:
+                  widget.beaches.map((beach) => beach.build(context)).toList(),
             ),
           ),
         ),
       );
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty<TransitionAnimation?>('progress', progress))
+      ..add(DiagnosticsProperty<OverlayState?>('overlay', overlay));
+  }
 }
 
 class _DeclarativeOverlay extends StatefulWidget {
   const _DeclarativeOverlay(
-      {@required this.child, @required this.overlayKey, Key key})
-      : assert(child != null),
-        assert(overlayKey != null),
-        super(key: key);
+      {required this.child, required this.overlayKey, Key? key})
+      : super(key: key);
 
   final Widget child;
   final Key overlayKey;
 
   @override
   _DeclarativeOverlayState createState() => _DeclarativeOverlayState();
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DiagnosticsProperty<Key>('overlayKey', overlayKey));
+  }
 }
 
 class _DeclarativeOverlayState extends State<_DeclarativeOverlay> {
-  OverlayEntry _childEntry;
+  late OverlayEntry _childEntry;
 
   @override
   void initState() {
@@ -234,7 +271,9 @@ class _DeclarativeOverlayState extends State<_DeclarativeOverlay> {
   @override
   void didUpdateWidget(_DeclarativeOverlay oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.child != widget.child) _childEntry.markNeedsBuild();
+    if (oldWidget.child != widget.child) {
+      _childEntry.markNeedsBuild();
+    }
   }
 
   @override
@@ -246,25 +285,25 @@ class _DeclarativeOverlayState extends State<_DeclarativeOverlay> {
 
 /// See [Coast]
 class Beach {
-  Beach({@required this.builder}) : assert(builder != null);
+  Beach({required this.builder});
 
   final WidgetBuilder builder;
 
   final GlobalKey _subtreeKey = GlobalKey();
 
-  BuildContext get subtreeContext => _subtreeKey.currentContext;
+  BuildContext? get subtreeContext => _subtreeKey.currentContext;
 
   Widget build(BuildContext context) =>
       RepaintBoundary(key: _subtreeKey, child: Builder(builder: builder));
 }
 
 class CoastObserver {
-  CoastState coast;
+  late CoastState coast;
 
   /// [progress] will animate between 0 and 1. It will always start at 0, and end either at 1 (when the transition
   /// completes) or at 0 (when the transition is cancelled).
   void startTransition(Beach beach, Beach previousBeach,
-      BeachTransitionDirection direction, Animation<double> progress) {}
+      BeachTransitionDirection direction, Animation<double>? progress) {}
 }
 
 enum BeachTransitionDirection { left, right }
@@ -274,11 +313,11 @@ class TransitionAnimation extends Animation<double>
         AnimationLocalListenersMixin,
         AnimationLocalStatusListenersMixin,
         AnimationEagerListenerMixin {
-  double _value = 0.0;
-  AnimationStatus _status = AnimationStatus.dismissed;
+  double _value = 0;
+  AnimationStatus? _status = AnimationStatus.dismissed;
 
   @override
-  AnimationStatus get status => _status;
+  AnimationStatus get status => _status!;
 
   @override
   double get value => _value;
@@ -293,7 +332,7 @@ class TransitionAnimation extends Animation<double>
   AnimationStatus _lastReportedStatus = AnimationStatus.dismissed;
 
   @visibleForTesting
-  AnimationStatus calculateStatus(double newValue) {
+  AnimationStatus? calculateStatus(double newValue) {
     if (newValue == 0.0) {
       return AnimationStatus.dismissed;
     } else if (newValue == 1.0) {
@@ -304,7 +343,7 @@ class TransitionAnimation extends Animation<double>
       return AnimationStatus.reverse;
     } else {
       //this can never happen but is a fallback scenario
-      return null;
+      return AnimationStatus.dismissed;
     }
   }
 
